+++ src/lib/browser/fingerprint.py
@@ -138,7 +138,7 @@ class Fingerprint(object):
     STATIC_CATEGORY = 'static'
     RUNTIME_CATEGORY = 'runtime'
     TECHNOLOGY_RUNTIME_MAP = {
-        'WordPress': 'PHP', 'WooCommerce': 'PHP', 'Drupal': 'PHP', 'Joomla': 'PHP', 'Magento': 'PHP', 'Bitrix': 'PHP', 'OpenCart': 'PHP', 'PrestaShop': 'PHP', 'TYPO3': 'PHP', 'Nextcloud': 'PHP', 'ownCloud': 'PHP', 'Matomo': 'PHP', 'phpMyAdmin': 'PHP', 'phpBB': 'PHP', 'Moodle': 'PHP', 'Open Journal Systems': 'PHP', 'InstantCMS': 'PHP', 'Laravel': 'PHP', 'Symfony': 'PHP', 'Craft CMS': 'PHP', 'Bolt CMS': 'PHP', 'RoundCube Webmail': 'PHP', 'WHMCS': 'PHP', 'CS-Cart': 'PHP', 'CubeCart': 'PHP', 'DataLife Engine': 'PHP', 'Discuz!': 'PHP', 'SilverStripe': 'PHP', 'Webasyst / Shop-Script': 'PHP', 'XOOPS': 'PHP', 'Zen Cart CMS': 'PHP', 'e107': 'PHP', 'phpWind': 'PHP', 'phpCMS': 'PHP',
+        'WordPress': 'PHP', 'WooCommerce': 'PHP', 'Drupal': 'PHP', 'Joomla': 'PHP', 'Magento': 'PHP', 'Bitrix': 'PHP', 'OpenCart': 'PHP', 'PrestaShop': 'PHP', 'TYPO3': 'PHP', 'Nextcloud': 'PHP', 'ownCloud': 'PHP', 'Matomo': 'PHP', 'phpMyAdmin': 'PHP', 'phpBB': 'PHP', 'Moodle': 'PHP', 'Open Journal Systems': 'PHP', 'InstantCMS': 'PHP', 'DiafanCMS': 'PHP', 'Laravel': 'PHP', 'Symfony': 'PHP', 'Craft CMS': 'PHP', 'Bolt CMS': 'PHP', 'RoundCube Webmail': 'PHP', 'WHMCS': 'PHP', 'CS-Cart': 'PHP', 'CubeCart': 'PHP', 'DataLife Engine': 'PHP', 'Discuz!': 'PHP', 'SilverStripe': 'PHP', 'Webasyst / Shop-Script': 'PHP', 'XOOPS': 'PHP', 'Zen Cart CMS': 'PHP', 'e107': 'PHP', 'phpWind': 'PHP', 'phpCMS': 'PHP',
         'Express': 'Node.js', 'NestJS': 'Node.js', 'Fastify': 'Node.js', 'Koa': 'Node.js', 'Hapi': 'Node.js', 'Strapi': 'Node.js', 'Directus': 'Node.js', 'Ghost': 'Node.js', 'Next.js': 'Node.js', 'Nuxt': 'Node.js', 'Gatsby': 'Node.js', 'Astro': 'Node.js', 'Remix': 'Node.js', 'SvelteKit': 'Node.js', 'Docusaurus': 'Node.js', 'VitePress': 'Node.js', 'PencilBlue': 'Node.js',
         'React': 'JavaScript', 'Vue': 'JavaScript', 'Angular': 'JavaScript', 'Django': 'Python', 'Flask': 'Python', 'FastAPI': 'Python', 'Ruby on Rails': 'Ruby', 'Spree': 'Ruby', 'Spring': 'Java/JVM', 'Liferay': 'Java/JVM', 'OpenCms': 'Java/JVM', 'Hippo CMS': 'Java/JVM', 'dotCMS': 'Java/JVM', 'ASP.NET': '.NET', 'Microsoft SharePoint': '.NET', 'DNN Platform': '.NET', 'Orchard CMS': '.NET', 'Sitecore': '.NET', 'Sitefinity': '.NET', 'Umbraco': '.NET', 'Phoenix': 'Elixir', 'MkDocs': 'Static site', 'Jekyll': 'Static site', 'Hugo': 'Static site', 'AsciiDoc': 'Static site',
     }
@@ -164,6 +164,7 @@ class Fingerprint(object):
         ('CS-Cart', ECOMMERCE_CATEGORY, ('cs-cart', 'cs cart', 'cscart')),
         ('CubeCart', ECOMMERCE_CATEGORY, ('cubecart',)),
         ('DataLife Engine', CMS_CATEGORY, ('datalife engine', 'dle')),
+        ('DiafanCMS', CMS_CATEGORY, ('diafan.cms', 'diafan cms')),
         ('Discuz!', CMS_CATEGORY, ('discuz!', 'discuz')),
         ('Duda', SITE_BUILDER_CATEGORY, ('duda', 'duda website builder')),
         ('DNN Platform', CMS_CATEGORY, ('dnn platform', 'dotnetnuke', 'dnn')),
@@ -240,6 +241,7 @@ class Fingerprint(object):
         ('CS-Cart', ECOMMERCE_CATEGORY, ('var tygh', 'index.php?dispatch=', '/design/themes/')),
         ('CubeCart', ECOMMERCE_CATEGORY, ('cubecart', 'index.php?_a=', '/skins/')),
         ('DataLife Engine', CMS_CATEGORY, ('datalife engine', 'engine/ajax/', 'index.php?do=')),
+        ('DiafanCMS', CMS_CATEGORY, ('diafan.cms',)),
         ('Discuz!', CMS_CATEGORY, ('discuz_uid', 'discuz_tips', 'static/image/common/', 'powered by discuz')),
         ('Duda', SITE_BUILDER_CATEGORY, ('static-cdn.multiscreensite.com', 'data-cmsid=', 'dmcdn.net')),
         ('DNN Platform', CMS_CATEGORY, ('__dnnvariable', 'dnn_', '/portals/_default/')),
@@ -1420,6 +1422,28 @@ class Fingerprint(object):
                     break
 
     @staticmethod
+    def _has_diafan_meta_author(body_lower):
+        """
+        Return True when an HTML meta author tag exposes DiafanCMS branding.
+
+        DiafanCMS commonly exposes a non-generator marker such as
+        <meta content="DIAFAN.CMS https://www.diafan.ru/" name="author">.
+        The check is constrained to meta tags to avoid classifying generic
+        body text that merely links to or mentions Diafan.
+
+        :param str body_lower: normalized response body
+        :return: bool
+        """
+
+        body_text = str(body_lower or '')
+        return (
+            re.search(r'<meta[^>]+name=["\']author["\'][^>]+content=["\'][^"\']*diafan\.cms', body_text)
+            is not None
+            or re.search(r'<meta[^>]+content=["\'][^"\']*diafan\.cms[^"\']*["\'][^>]+name=["\']author["\']', body_text)
+            is not None
+        )
+
+    @staticmethod
     def _header_contains(headers, name, needle):
         """
         Case-insensitive header contains check.
@@ -1708,6 +1732,10 @@ class Fingerprint(object):
             self._add_signal('Shopify', self.ECOMMERCE_CATEGORY, 'cookie', '_shopify*', 6)
         if 'x-shopid' in headers or 'shopify' in server:
             self._add_signal('Shopify', self.ECOMMERCE_CATEGORY, 'header', 'x-shopid|server', 6)
+
+        # DiafanCMS
+        if self._has_diafan_meta_author(body_lower):
+            self._add_signal('DiafanCMS', self.CMS_CATEGORY, 'meta', 'author=DIAFAN.CMS', 8)
 
         # Bitrix
         if 'bitrix' in x_powered_cms: